Line 10:
I feel like a contrast sets up a better delivery for the thesis.

Line 11:

It is versus it's: I feel like the full form is better when used outside of speech. For example, on lines 73 and 74 the text reads more like speech. I like this distinction in formality and feel it will sharpen/justify the informal bits more distinctly.